backend/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
-e git+https://github.com/django-tastypie/django-tastypie.git@ac18543860a1deb88ed83c6d2eecad02a525419d#egg=django-tastypie
    # via -r requirements.in
asgiref==3.4.1
    # via django
cachetools==4.1.0
    # via google-auth
certifi==2020.6.20
    # via
    #   requests
    #   sentry-sdk
chardet==3.0.4
    # via requests
cycler==0.10.0
    # via matplotlib
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.8
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==3.10.0
    # via -r requirements.in
djangorestframework==3.12.4
    # via -r requirements.in
google-api-core==1.21.0
    # via google-api-python-client
google-api-python-client==1.9.3
    # via -r requirements.in
google-auth==1.18.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.0.3
    # via google-api-python-client
googleapis-common-protos==1.52.0
    # via google-api-core
gunicorn==20.1.0
    # via -r requirements.in
httplib2==0.19.1
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==2.9
    # via requests
kiwisolver==1.2.0
    # via matplotlib
lxml==4.6.3
    # via -r requirements.in
matplotlib==3.3.0
    # via -r requirements.in
numpy==1.20.3
    # via matplotlib
pika==1.1.0
    # via -r requirements.in
pillow==8.3.2
    # via
    #   -r requirements.in
    #   matplotlib
protobuf==3.12.2
    # via
    #   google-api-core
    #   googleapis-common-protos
psycopg2==2.9.1
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pyparsing==2.4.7
    # via
    #   httplib2
    #   matplotlib
python-dateutil==2.8.0
    # via
    #   django-tastypie
    #   matplotlib
python-mimeparse==1.6.0
    # via django-tastypie
pytz==2019.1
    # via
    #   django
    #   google-api-core
requests==2.25.1
    # via google-api-core
rsa==4.7.2
    # via google-auth
sentry-sdk==1.3.0
    # via -r requirements.in
shortuuid==1.0.1
    # via -r requirements.in
simplejson==3.16.0
    # via -r requirements.in
six==1.12.0
    # via
    #   cycler
    #   google-api-core
    #   google-api-python-client
    #   google-auth
    #   protobuf
    #   python-dateutil
sqlparse==0.3.0
    # via django
uritemplate==3.0.1
    # via google-api-python-client
urllib3==1.26.5
    # via
    #   requests
    #   sentry-sdk

# The following packages are considered to be unsafe in a requirements file:
# setuptools
